Adds a confirm dialog to restart PC

+++ auto_update.py
@@ -1,8 +1,12 @@
 # Windows Auto Update script
-# VERSION: 1.3.0
+# VERSION: 1.4.0
 # AUTHOR: Valentin Le Gal
 
-import os, shutil
+
+import os
+from pyautogui import confirm
+from threading import Thread
+from time import sleep
 
 
 class Computer:
@@ -10,6 +14,7 @@ class Computer:
 		self.desktop_private_path = os.path.join(os.path.join(os.environ["USERPROFILE"]), "Desktop")
 		self.desktop_public_path = "C:\\Users\\Public\\Desktop"
 		self.saved_files = []
+		self.thread_running = False
 
 	def save_folder_files(self, folder):
 		for item in os.listdir(folder):
@@ -19,29 +24,23 @@ class Computer:
 		for item in os.listdir(folder):
 			if not item in self.saved_files:
 				path = os.path.join(folder, item)
-				
+
 				# Delete these files
 				try:
-					if os.path.exists(path):
-						if os.path.isdir(path):
-							if os.path.islink(path):
-								os.unlink(path)
-							else:
-								shutil.rmtree(path)
-						else:
-							if os.path.islink(path):
-								os.unlink(path)
-							else:
-								os.remove(path)
+					if os.path.islink(path):
+						os.unlink(path)
+					elif os.path.isfile(path):
+						os.remove(path)
+
 				except Exception as error:
-					print("[ERROR] Delete files: " + error)
+					print("[ERROR] Delete links: " + error)
 	
 	def create_restore_point(self, name, type):
 		os.system("powershell.exe Checkpoint-Computer -Description \"" + name + "\" -RestorePointType \"" + type + "\"")
 		os.system("cls")
 
 	def choco_update(self):
-		os.system("choco upgrade all -y")
+		os.system("choco upgrade all -y --ignore-checksums")
 		os.system("cls")
 
 	def windows_update(self):
@@ -51,8 +50,34 @@ class Computer:
 		os.system("cls")
 
 	def restart(self):
-		os.system("powershell.exe Restart-Computer -Force")
+		self.thread_running = True
+		
+		t1 = Thread(target=self.dialog)
+		t2 = Thread(target=self.timer)
 
+		t1.start()
+		t2.start()
+
+	def dialog(self):
+		message = "Votre PC va redémarrer dans environ 60 secondes...\nVoulez-vous le redémarrer maintenant ?"
+		response = confirm(message, "Windows Auto Update")
+
+		if response == "OK":
+			self.thread_running = False
+			print("PC will restart")
+			os.system("powershell.exe Restart-Computer -Force")
+		else:
+			self.thread_running = False
+
+	def timer(self):
+		cpt = 0
+		while self.thread_running and cpt < 60:
+			sleep(1)
+			cpt = cpt + 1
+		
+		if cpt >= 60:
+			self.thread_running = False
+			os.system("powershell.exe Restart-Computer -Force")
 
 # Create the Computer object
 pc = Computer()
@@ -74,5 +99,5 @@ pc.check_new_folder_files(pc.desktop_public_path)
 # Updates Windows with PSWindowsUpdate
 pc.windows_update()
 
-# Restart the computer
+# Restart the computer (with Confirm Dialog)
 pc.restart()
